Add files via upload
Description:
FoodieHub is a restaurant clone project that I developed using HTML, CSS, and JavaScript. This project aims to replicate the user interface and functionalities of a restaurant website, allowing users to explore the menu, make reservations, and get in touch with the restaurant.

Technology used:
The project is built using core web technologies, including:

HTML5: For structuring the content of the web pages.
CSS3: For styling and layout, ensuring an appealing visual design.
JavaScript: To implement interactivity and dynamic elements, enhancing the user experience.